remove presentation page, instead connect onFindChannel to the optimizer page
